LayoutTests/editing/selection/move-to-line-boundary-clear-selection-expected.txt

PASS getSelection().toString() is ''
PASS getSelection().toString() is ''
webkit
